refactor(mamba2): split Mamba2Block into base + subclasses (Option B+D)
Split the monolithic 1047-line Mamba2Block class into:

- _mamba_block.py: Mamba2BlockBase (shared __init__), Mamba2BlockSingle
  (single-token path), and Mamba2Block factory function
- _mamba_block_scan.py: Mamba2BlockScan (ONNX Scan token-by-token)
- _mamba_block_chunked.py: Mamba2BlockChunkedSSD (parallel chunked SSD)

The factory function reads flags.mamba_scan at construction time and
returns the matching subclass. Zero caller changes needed — all 4
consumers (bamba, granitemoehybrid, mamba, nemotron_h) use the factory
exactly like the old class.

Design follows my/mamba-2-design.md Option B+D:
- Subclass over boolean flags (no mode dispatch in forward())
- _realize_submodule is a single staticmethod on Mamba2BlockBase
- Chunked SSD helpers (_segment_sum, _segment_sum_dynamic) move to
  the chunked file
